Prepare java-client to run in CI under the mock.
Note that this changeset touches a lot of files which do not work correctly
under the mock either because the code is too tied towards the actual cluster
management or the mock just plainly does not support it.

We still run lots of tests though, so I deemed its better to run those and
ignored the rest, which we can revisit test by test and rework to reenable
if needed.

Reviewed-on: http://review.couchbase.org/101905
Tested-by: Build Bot <build@couchbase.com>

+++ src/integration/java/com/couchbase/client/java/N1qlClusterLevelTest.java
@@ -17,6 +17,7 @@ package com.couchbase.client.java;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
+import static org.junit.Assume.assumeFalse;
 import static org.junit.Assume.assumeTrue;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyString;
@@ -60,6 +61,8 @@ public class N1qlClusterLevelTest {
 
     @BeforeClass
     public static void init() {
+        assumeFalse(CouchbaseTestContext.isMockEnabled());
+
         ctx = CouchbaseTestContext.builder()
                 .bucketName("N1qlCluster")
                 .bucketPassword("protected")
@@ -103,9 +106,15 @@ public class N1qlClusterLevelTest {
 
     @AfterClass
     public static void cleanup() {
-        ctx.destroyBucketAndDisconnect();
-        ctx2.destroyBucketAndDisconnect();
-        ctx3.destroyBucketAndDisconnect();
+        if (ctx != null) {
+            ctx.destroyBucketAndDisconnect();
+        }
+        if (ctx2 != null) {
+            ctx2.destroyBucketAndDisconnect();
+        }
+        if (ctx3 != null) {
+            ctx3.destroyBucketAndDisconnect();
+        }
     }
 
     @Test
